… cub/detail/strong_store.cuh